Updated gradle version, Android SDK, .gitignore, kotlin version
