+++ .gitignore
@@ -14,18 +14,9 @@ local.properties
 .signing/
 
 # User-specific configurations
-.idea/libraries/
-.idea/workspace.xml
-.idea/tasks.xml
-.idea/.name
-.idea/compiler.xml
-.idea/copyright/profiles_settings.xml
-.idea/encodings.xml
-.idea/misc.xml
-.idea/modules.xml
-.idea/scopes/scope_settings.xml
-.idea/vcs.xml
+/.idea/
 *.iml
+/build/
 
 # OS-specific files
 .DS_Store
@@ -35,3 +26,5 @@ local.properties
 .Trashes
 ehthumbs.db
 Thumbs.db
+
+
