+++ build.gradle
@@ -1,12 +1,12 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.0.5-2'
+    ext.kotlin_version = '1.1.0'
     repositories {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.2'
+        classpath 'com.android.tools.build:gradle:2.3.0'
         classpath 'com.novoda:bintray-release:0.3.4'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath 'com.github.dcendents:android-maven-gradle-plugin:1.5'
@@ -19,6 +19,15 @@ allprojects {
     repositories {
         jcenter()
     }
+}
+
+ext {
+    compileSdk = 25
+    buildTools = "25.0.2"
+    minSdk = 14
+    targetSdk = 25
+
+    supportLibrary = "25.2.0"
 }
 
 task clean(type: Delete) {
