+++ fingerprintauthhelper/build.gradle
@@ -4,15 +4,14 @@ apply plugin: 'kotlin-android'
 apply plugin: 'com.github.dcendents.android-maven'
 
 group='com.github.pro100svitlo'
-version = '1.2.1'
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.1"
+    compileSdkVersion project.compileSdk
+    buildToolsVersion project.buildTools
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 24
+        minSdkVersion project.minSdk
+        targetSdkVersion project.targetSdk
         versionCode 1
         versionName "1.0"
     }
@@ -39,7 +38,7 @@ task javadoc(type: Javadoc) {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:24.2.1'
+    compile "com.android.support:appcompat-v7:$supportLibrary"
     compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 }
 
@@ -48,7 +47,7 @@ dependencies {
 publish {
     groupId = 'com.github.pro100svitlo'
     artifactId = 'fingerprintAuthHelper'
-    publishVersion = '1.2.1'
+    publishVersion = '1.2.2'
     desc = 'A small library that allow You easy manage fingerprint authentication on devices with ' +
             'fingerprint scanner on Android M and upper'
     licences = ['MIT']
