+++ sample/build.gradle
@@ -2,16 +2,15 @@ apply plugin: 'com.android.application'
 apply plugin: 'com.github.dcendents.android-maven'
 
 group='com.github.pro100svitlo'
-version = '1.0.3'
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.2"
+    compileSdkVersion project.compileSdk
+    buildToolsVersion project.buildTools
 
     defaultConfig {
         applicationId "com.pro100svitlo.fingerprintauthdemo"
-        minSdkVersion 14
-        targetSdkVersion 24
+        minSdkVersion project.minSdk
+        targetSdkVersion project.targetSdk
         versionCode 4
         versionName "1.0.3"
     }
@@ -29,8 +28,8 @@ android {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:24.2.1'
-    compile 'com.android.support:design:24.2.1'
+    compile "com.android.support:appcompat-v7:$supportLibrary"
+    compile "com.android.support:design:$supportLibrary"
 //    compile 'com.github.pro100svitlo:fingerprintAuthHelper:1.1.9'
     compile project(path: ':fingerprintauthhelper')
 }
